12 - Capstone RAG/requirements.txt
gradio
requests
openai
PyPDF2
python-dotenv
faiss-cpu
pydantic
langchain-community
langchain-openai
tiktoken